Introduce `AbstractFileSystem`
